[Feat #26] 관심목록 조회

+++ src/main/java/JJinBBang/app/domain/building/dto/BasicInfo.java
@@ -1,8 +1,9 @@
 package JJinBBang.app.domain.building.dto;
 
 import JJinBBang.app.domain.building.entity.Buildings;
-import JJinBBang.app.domain.building.enums.BuildingType;
 import lombok.Builder;
+
+import java.math.BigDecimal;
 
 @Builder
 public record BasicInfo(
@@ -11,7 +12,7 @@ public record BasicInfo(
 	String type,
 	String name,
 	String address,
-	Double rating,
+	BigDecimal rating,
 	Integer reviewCount
 ) {
 	public static BasicInfo of(Buildings building, Boolean liked) {
